Change `TileExcelImporter` ancestor to `BaseImportModule`
Move some methods from BranchExcelImporter to TileExcelImporter
